requirements.txt
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
huggingface-hub==0.20.0
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
